+++ src/pages/pmf-game/index.tsx
@@ -3,10 +3,10 @@ import SEO from 'components/seo'
 import ScrollArea from 'components/RadixUI/ScrollArea'
 import { GameState, GameActions, SceneId, Resource, SavedResource, ChecklistItem } from '../../data/pmf-game/types'
 import { LEVELS, getLevelById, getInitialLevelProgress } from '../../data/pmf-game/levels'
-import TitleScene from './components/TitleScene'
-import Sidenav from './components/Sidenav'
-import OverworldScene from './components/OverworldScene'
-import LevelScene from './components/LevelScene'
+import TitleScene from 'components/pmf-game/TitleScene'
+import Sidenav from 'components/pmf-game/Sidenav'
+import OverworldScene from 'components/pmf-game/OverworldScene'
+import LevelScene from 'components/pmf-game/LevelScene'
 
 const STORAGE_KEY = 'posthog-pmf-game-state'
